Fix bug with ParentWrapper::get returning unescaped HTML values
In the case of a ParentWrapper being passed an expression, it was
incorrectly using this.expression.get(), which does not support
escaping values. Thus, this didn't escape HTML in certain cases.

Fixing this bug might break UIs that previously relied on the value not
being escaped properly, but because this is a bug fix and the prior
behavior was incorrect, it will be released as a patch version.