implement DependencyOptions & ignoreTemplate option
Switching dependencies to take a second options argument. The ignoreTemplate option will be used for avoiding looking for dependencies within a given template.

This is useful when a template is going to be rendered separately, since it will emit its own dependencies. This happens in the case of a DynamicText whose value ends up being a template. In this case, the template itself is rendered in the postiion of the dynamic text, and we don't need to look for dependencies within the template at the time of rendering the DynamicText.