remove code that iterates through returned values to render templates within them
This was implmented in particular to handle rendering the values created by array template tags. This was always pretty dodgy and we should more ideally be using expressions that return the rendered or non rendered template value as appropriate. We'll replace this with new expression types specifically for creating arrays and objects rather than relying on the general operator implementation