TemplateClosure => ContextClosure
Rethink TemplateClosure, which didn't work in the case of `within` attributes. Rather than switching out the context for a fixed reference, create a child context, more like how other types of context lookups work. Only use the closure when the lookup is happening as a result of a ViewParent.

This fixes buggy behavior when using `within` on an attribute. Using `within` doesn't create a ViewParent, so the context is not altered.