implement DeferRenderExpression
This Expression is used to wrap a template so that when its containing Expression--such as an ObjectExpression or ArrayExpression--is evaluated, it returns the template unrendered and wrapped in the current context. Separating evaluation of the containing expression from template rendering is used to support array attributes of views. This way, we can evaluate an array and iterate through it separately from rendering template content