always wrap templates in ContextClosure when returned from get method and not wrapped already
It could be possible for an object to have an unwrapped Template in it that is returned by looking up segemtns, so this shouldn't be an else.

Also, wrapping a template in two ContextClosures isn't needed, since the inner ContextClosure is going to be the only one that has an effect. Thus, do check for an existing ContextClosure to avoid direct double wrapping.