implement equals method for equivalent ContextClosure objects (#20)
* implement equals method for equivalent ContextClosure objects

* implement DependencyOptions::setIgnoreTemplate() & use ContextClosure source template

The desire of ignoreTemplate is to prevent over-computing bindings for templates. However, wrapping templates in ContextClosures was preventing us from detecting the template that should be ignored. This resolves the issue by ignoring the base template of the ContextClosure

* always wrap templates in ContextClosure when returned from get method and not wrapped already

It could be possible for an object to have an unwrapped Template in it that is returned by looking up segemtns, so this shouldn't be an else.

Also, wrapping a template in two ContextClosures isn't needed, since the inner ContextClosure is going to be the only one that has an effect. Thus, do check for an existing ContextClosure to avoid direct double wrapping.

* more efficient ignore that uses contextClosure.template

* use template.equals to handle possible recursive ContextClosure equality checks